+++ llm_relay_desk/proxy/__init__.py
@@ -1,6 +1,7 @@
 from .common import (
     error_from_body,
     native_upstream_root,
+    openai_upstream_base,
     timeout_config,
     upstream_headers,
     verify_local_api_key,
@@ -12,7 +13,12 @@ __all__ = [
     "OpenAISSEParser",
     "error_from_body",
     "native_upstream_root",
+    "openai_upstream_base",
+    "configured_upstream_protocol",
+    "resolve_upstream_protocol",
     "timeout_config",
     "upstream_headers",
     "verify_local_api_key",
 ]
+
+from .protocol import configured_upstream_protocol, resolve_upstream_protocol